+++ internal-api/src/main/java/datadog/trace/api/telemetry/IntegrationsCollector.java
@@ -1,4 +1,4 @@
-package datadog.trace.api;
+package datadog.trace.api.telemetry;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
