+++ telemetry/src/main/java/datadog/telemetry/integration/IntegrationPeriodicAction.java
@@ -3,7 +3,7 @@ package datadog.telemetry.integration;
 import datadog.telemetry.TelemetryRunnable;
 import datadog.telemetry.TelemetryService;
 import datadog.telemetry.api.Integration;
-import datadog.trace.api.IntegrationsCollector;
+import datadog.trace.api.telemetry.IntegrationsCollector;
 import java.util.Map;
 
 public class IntegrationPeriodicAction implements TelemetryRunnable.TelemetryPeriodicAction {
